Move discounts methods to separate utils file
